Move about and share button to components folder
